*** empty log message ***
svn path=/trunk/boinc/; revision=6602

+++ client/hostinfo_unix.C
@@ -420,7 +420,7 @@ int HOST_INFO::get_host_info() {
 #elif defined(_HPUX_SOURCE)
     struct pst_static pst; 
     pstat_getstatic(&pst, sizeof(pst), (size_t)1, 0);
-    m_nbytes = (pst.physical_memory * pst.page_size);
+    m_nbytes = (double)pst.physical_memory * (double)pst.page_size;
 #else
 #error Need to specify a sysconf() define to obtain memory size
 #endif
